+++ LibRbxl/UDim2.cs
@@ -11,9 +11,9 @@ namespace LibRbxl
         public UDim X { get; }
         public UDim Y { get; }
 
-        public float OffsetX => X.Offset;
+        public int OffsetX => X.Offset;
         public float ScaleX => X.Scale;
-        public float OffsetY => Y.Offset;
+        public int OffsetY => Y.Offset;
         public float ScaleY => Y.Scale;
 
         public UDim2(UDim x, UDim y)
@@ -22,7 +22,7 @@ namespace LibRbxl
             Y = y;
         }
 
-        public UDim2(float offsetX, float scaleX, float offsetY, float scaleY)
+        public UDim2(int offsetX, float scaleX, int offsetY, float scaleY)
         {
             X = new UDim(offsetX, scaleX);
             Y = new UDim(offsetY, scaleY);
